+++ CHANGELOG.md
@@ -13,6 +13,7 @@ When updating the changelog, remember to be very clear about what behavior has c
 and what APIs have changed, if applicable.
 
 ## [Unreleased]
+- Handle HTTP/2 stream errors in tracker client and LB state updator.
 
 ## [29.55.0] - 2024-05-23
 - Allow HttpBridge to return RetriableRequestException for the Netty max active stream error